feat: improvements for ai flow, bug fixes
