+++ src/components/App/SideBar/AiSummary/AiSources/index.tsx
@@ -1,3 +1,4 @@
+import { Slide } from '@mui/material'
 import Button from '@mui/material/Button'
 import { memo, useCallback, useRef, useState } from 'react'
 import styled from 'styled-components'
@@ -41,19 +42,21 @@ const _AiSources = ({ sourceIds }: Props) => {
 
   return (
     <SectionWrapper>
-      <Heading align="center" className="heading" direction="row" justify="space-between">
-        <Flex align="center" direction="row">
-          <div className="heading__icon">
-            <SourcesIcon />
-          </div>
-          <span className="tittle">Sources</span>
-          <span className="heading__count">{sourceIds.length}</span>
-        </Flex>
-        <CollapseButton onClick={handleLoadMoreClick}>
-          {showAll ? 'Hide all' : 'Show all'}
-          {showAll ? <ChevronUpIcon /> : <ChevronDownIcon />}
-        </CollapseButton>
-      </Heading>
+      <Slide direction="right" in mountOnEnter>
+        <Heading align="center" className="heading" direction="row" justify="space-between">
+          <Flex align="center" direction="row">
+            <div className="heading__icon">
+              <SourcesIcon />
+            </div>
+            <span className="tittle">Sources</span>
+            <span className="heading__count">{sourceIds.length}</span>
+          </Flex>
+          <CollapseButton onClick={handleLoadMoreClick}>
+            {showAll ? 'Hide all' : 'Show all'}
+            {showAll ? <ChevronUpIcon /> : <ChevronDownIcon />}
+          </CollapseButton>
+        </Heading>
+      </Slide>
       {showAll && visibleNodes.length > 0 && (
         <ScrollView ref={scrollViewRef} id="search-result-list" shrink={1}>
           {visibleNodes.map((n, index) => {
