+++ src/stores/useAiSummaryStore/index.ts
@@ -11,15 +11,18 @@ export type AiSummaryStore = {
   aiSummaryAnswers: AIAnswer
   aiRefId: string
   setAiSummaryAnswer: (key: string, answer: AIEntity) => void
+  setNewLoading: (answer: AIEntity | null) => void
   resetAiSummaryAnswer: () => void
   getAiSummaryAnswer: (key: string) => string
   getKeyExist: (key: string) => boolean
   setAiRefId: (aiRefId: string) => void
+  newLoading: AIEntity | null
 }
 
 const defaultData = {
   aiSummaryAnswers: {},
   aiRefId: '',
+  newLoading: null,
 }
 
 export const useAiSummaryStore = create<AiSummaryStore>()(
@@ -33,6 +36,9 @@ export const useAiSummaryStore = create<AiSummaryStore>()(
       const clone = structuredClone(summaryAnswers)
 
       set({ aiSummaryAnswers: clone })
+    },
+    setNewLoading: (newLoading) => {
+      set({ newLoading })
     },
     resetAiSummaryAnswer: () => {
       set({ aiSummaryAnswers: {}, aiRefId: '' })
@@ -55,11 +61,11 @@ export const useAiSummaryStore = create<AiSummaryStore>()(
   })),
 )
 
-export const useHasAiChats = () => useAiSummaryStore((s) => !isEmpty(s.aiSummaryAnswers))
+export const useHasAiChats = () => useAiSummaryStore((s) => !isEmpty(s.aiSummaryAnswers) || !!s.newLoading)
 
-export const useHasAiChatsResponse = () =>
+export const useHasAiChatsResponseLoading = () =>
   useAiSummaryStore((s) => {
     const answers = s.aiSummaryAnswers
 
-    return Object.values(answers).at(-1)?.answerLoading
+    return !!s.newLoading || Object.values(answers).at(-1)?.answerLoading
   })
