+++ src/stores/useDataStore/index.ts
@@ -137,7 +137,7 @@ export const useDataStore = create<DataStore>()(
     fetchData: async (setBudget, setAbortRequests, AISearchQuery = '') => {
       const { currentPage, itemsPerPage, dataInitial: existingData, filters } = get()
       const { currentSearch } = useAppStore.getState()
-      const { setAiSummaryAnswer, aiRefId } = useAiSummaryStore.getState()
+      const { setAiSummaryAnswer, setNewLoading, aiRefId } = useAiSummaryStore.getState()
       let ai = { ai_summary: String(!!AISearchQuery) }
 
       if (!AISearchQuery) {
@@ -150,6 +150,7 @@ export const useDataStore = create<DataStore>()(
 
       if (AISearchQuery) {
         ai = { ...ai, ai_summary: String(true) }
+        setNewLoading({ question: AISearchQuery, answerLoading: true })
       }
 
       if (abortController) {
@@ -195,6 +196,8 @@ export const useDataStore = create<DataStore>()(
             sourcesLoading: !answer,
             shouldRender: true,
           })
+
+          setNewLoading(null)
         }
 
         const currentNodes = currentPage === 0 && !aiRefId ? [] : [...(existingData?.nodes || [])]
